Autowire and Dependency Injection
First two classes were created (dev, device). Exception null point error was seen. We fixed it by putting @Component in our device.java file. So now our spring container has 2 classes but the link between 2 classes aren't formed yet. device.java is dependent on dev.java to fix this we put @Autowired so that the object the device.java(d1) is instantiated and can be found BY TYPE (type as in dev or device). To find object BY NAME we use @Qualifier("name_of_the_object_we_seaching_for"). If we are to shorten the name of the object we can put it in the parameter of Qualifier and Component of that file we are searching